Added usercontrols for contentcontrol in mainwindow
