+++ TheMovies/App.xaml
@@ -12,13 +12,9 @@
             </ResourceDictionary.MergedDictionaries>
 
             <DataTemplate DataType="{x:Type viewmodel:HomeViewModel}">
-                <local:HomeView/>
+                <view:HomeView/>
             </DataTemplate>
-
-            <DataTemplate DataType="{x:Type viewmodel:MovieViewModel}">
-                <view:MovieView/>
-            </DataTemplate>
-
+            
             <DataTemplate DataType="{x:Type viewmodel:MovieProgramViewModel}">
                 <view:MovieProgramView/>
             </DataTemplate>
