+++ TheMovies/MVVM/View/BookingView.xaml
@@ -1,18 +1,16 @@
-﻿<Window x:Class="TheMovies.MVVM.View.BookingView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TheMovies.MVVM.View"
-        xmlns:ViewModel="clr-namespace:TheMovies.MVVM.ViewModel"
-        mc:Ignorable="d"
-        Title="BookingView" Height="900" Width="700"
-        Background="#272537">
-    <!-- shift + ctrl + b = Build uden at køre -->  
+﻿<UserControl x:Class="TheMovies.MVVM.View.BookingView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:TheMovies.MVVM.View"
+             xmlns:viewModel="clr-namespace:TheMovies.MVVM.ViewModel"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
     
-    <Window.DataContext>
-        <ViewModel:BookingViewModel/>
-    </Window.DataContext> 
+    <UserControl.DataContext>
+        <viewModel:BookingViewModel/>
+    </UserControl.DataContext>
 
     <Grid>
         <!-- Definerer rækker -->
@@ -24,7 +22,7 @@
         <!-- Definerer kolonner -->
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="45*"/>
-            <ColumnDefinition Width="45*"/> 
+            <ColumnDefinition Width="45*"/>
         </Grid.ColumnDefinitions>
 
         <!-- Overskrift -->
@@ -41,7 +39,7 @@
                       Height="50"
                       Width="300"
                       HorizontalAlignment="Right">
-                      
+
                 <TextBox Text="{Binding SearchTerm, UpdateSourceTrigger=PropertyChanged}"
                          Background="Transparent"
                          Foreground="AliceBlue"
@@ -50,7 +48,7 @@
             </GroupBox>
         </StackPanel>
 
-        
+
         <StackPanel Grid.Row="1" Orientation="Horizontal">
             <StackPanel>
                 <!-- Felt til angivelse af billetantal -->
@@ -227,16 +225,16 @@
                                        Background="AliceBlue"
                                        Margin="10 0 10 0"/>
 
-                                <Button Content="Tilføj"
+                            <Button Content="Tilføj"
                                         Command="{Binding AddBookingCommand,
                                         UpdateSourceTrigger=PropertyChanged}"
                                         Margin="10" Width="100"
                                         VerticalAlignment="Center"/>
-                            </StackPanel>
+                        </StackPanel>
                     </GroupBox>
-                    
+
                 </StackPanel>
-                
+
             </GroupBox>
 
             <GroupBox Header="Opdater/Fjern booking" Margin="10 5 0 0"
@@ -302,6 +300,6 @@
             </GroupBox>
 
         </StackPanel>
-        
+
     </Grid>
-</Window>
+</UserControl>
