+++ TheMovies/MVVM/View/HomeView.xaml
@@ -1,0 +1,15 @@
+﻿<UserControl x:Class="TheMovies.MVVM.View.HomeView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:TheMovies.MVVM.View"
+             xmlns:viewModel="clr-namespace:TheMovies.MVVM.ViewModel"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800"
+             Background="AliceBlue">
+
+    <Grid>
+        <Label Content="HOMEPAGE" HorizontalAlignment="Center" Margin="0,196,0,0" VerticalAlignment="Top" Foreground="black" FontSize="22"/>
+    </Grid>
+</UserControl>
