+++ TheMovies/MVVM/View/HomeView.xaml.cs
@@ -10,16 +10,17 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
 using System.Windows.Shapes;
 
 namespace TheMovies.MVVM.View
 {
     /// <summary>
-    /// Interaction logic for MainWindow1.xaml
+    /// Interaction logic for HomeView.xaml
     /// </summary>
-    public partial class MainWindow1 : Window
+    public partial class HomeView : UserControl
     {
-        public MainWindow1()
+        public HomeView()
         {
             InitializeComponent();
         }
