--- TheMovies/MVVM/View/MainWindow1.xaml
@@ -1,90 +1,0 @@
-﻿<Window x:Class="TheMovies.MVVM.View.MainWindow1"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TheMovies.MVVM.View"
-        xmlns:ViewModel="clr-namespace:TheMovies.MVVM.ViewModel"
-        mc:Ignorable="d"
-        Height="600" Width="910"
-        WindowStyle="None"
-        ResizeMode="NoResize"
-        Background="Transparent"
-        AllowsTransparency="True">
-    
-    <Window.DataContext>
-        <ViewModel:MainWindowViewModel />
-    </Window.DataContext>
-
-    <Border Background="#272537"
-        CornerRadius="20">
-        <Grid>
-
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="200"/>
-                <ColumnDefinition/>
-            </Grid.ColumnDefinitions>
-
-            <Grid.RowDefinitions>
-                <RowDefinition Height="75"/>
-                <RowDefinition/>
-            </Grid.RowDefinitions>
-
-            <!-- Knap lukker programmet -->
-            <Button Style="{DynamicResource ButtonStyle1}" Grid.Column="1" 
-                    Background="Transparent"
-                    Foreground="White"
-                    BorderBrush="Transparent"
-                    Content = "X"
-                    FontSize="16"
-                    FontWeight="Light"
-                    HorizontalAlignment="Right"
-                    Margin="0,0,20,0"
-                    Command="{Binding CloseMainWindowCommand}" />
-
-            <TextBlock Text="The Movies"
-               VerticalAlignment="Center"
-               HorizontalAlignment="Left"
-               Foreground="White"
-               FontSize="22"
-               Margin="20,0,0,0"/>
-
-            <StackPanel Grid.Row="1">
-
-                <RadioButton Content="Forside"
-                             Height="50"
-                             Foreground="White"
-                             FontSize="14"
-                             Style="{StaticResource MenuButtonTheme}"
-                             IsChecked="True"
-                             Command="{Binding HomeViewCommand}"/>
-
-                <RadioButton Content="Film"
-                             Height="50"
-                             Foreground="White"
-                             FontSize="14"
-                             Style="{StaticResource MenuButtonTheme}"
-                             Command="{Binding MovieViewCommand}"/>
-
-                <RadioButton Content="Programmer"
-                             Height="50"
-                             Foreground="White"
-                             FontSize="14"
-                             Style="{StaticResource MenuButtonTheme}"
-                             Command="{Binding MovieProgramViewCommand}"/>
-
-                <RadioButton Content="Bookinger"
-                             Height="50"
-                             Foreground="White"
-                             FontSize="14"
-                             Style="{StaticResource MenuButtonTheme}"
-                             Command="{Binding BookingViewCommand}"/>
-            </StackPanel>
-
-            <ContentControl Grid.Row="1"
-                            Grid.Column="1"
-                            Margin="10"
-                            Content="{Binding CurrentView}"/>
-        </Grid>
-    </Border>
-</Window>
