+++ TheMovies/MVVM/View/MovieProgramView.xaml.cs
@@ -1,24 +1,28 @@
-﻿using System.IO;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
 using System.Windows.Documents;
-using System.Windows.Xps.Packaging;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
 
 namespace TheMovies.MVVM.View
 {
     /// <summary>
-    /// Interaction logic for MovieProgram.xaml
+    /// Interaction logic for MovieProgramView.xaml
     /// </summary>
-    public partial class MovieProgramView : Window
+    public partial class MovieProgramView : UserControl
     {
         public MovieProgramView()
         {
             InitializeComponent();
-        }
-
-        private void ComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-
         }
     }
 }
