+++ TheMovies/MVVM/View/MovieView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace TheMovies.MVVM.View
+{
+    /// <summary>
+    /// Interaction logic for MovieView.xaml
+    /// </summary>
+    public partial class MovieView : UserControl
+    {
+        public MovieView()
+        {
+            InitializeComponent();
+        }
+    }
+}
