+++ TheMovies/MVVM/ViewModel/HomeViewModel.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections.ObjectModel;
+using System.ComponentModel;
+using System.Windows;
+using System.Windows.Data;
+using System.Windows.Input;
+using TheMovies.MVVM.Model.Classes;
+using TheMovies.MVVM.Model.Repositories;
+
+namespace TheMovies.MVVM.ViewModel
+{
+    public class HomeViewModel : ViewModelBase
+    {
+        
+        public HomeViewModel()
+        {
+
+        }
+    }
+
+}
